requirements.txt
streamlit==1.31.0
Pillow==10.0.0
pandas==2.0.3
easyocr==1.7.1
numpy==1.24.3